requirements.txt
google-auth
google-auth-oauthlib
google-api-python-client
google-auth-httplib2
streamlit
openai
langchain
tiktoken
python-dotenv

